Make package PSR-4 compatible and package:discover ready
