+++ .travis.yml
@@ -1,8 +1,10 @@
 language: php
 
-php: 
-  - 5.3
+php:
   - 5.4
+  - 5.6
+  - 7.0
+  - 7.1
 
 before_script:
   - curl -s http://getcomposer.org/installer | php
